+++ packages/yourloops/test/integration/pages/trends/trends-view.patient.spec.tsx
@@ -31,11 +31,11 @@ import { checkPatientNavBarAsPatient } from '../../assert/patient-nav-bar.assert
 import { minimalTrendViewData, mockDataAPI } from '../../mock/data.api.mock'
 import { renderPage } from '../../utils/render'
 import { checkPatientLayout } from '../../assert/layout.assert'
-import { unmonitoredPatientAsTeamMember } from '../../data/patient.api.data'
+import { patient2AsTeamMember } from '../../data/patient.api.data'
 
 describe('Trends view for patient', () => {
   beforeAll(() => {
-    mockPatientLogin(unmonitoredPatientAsTeamMember)
+    mockPatientLogin(patient2AsTeamMember)
   })
 
   it('should render correct layout', async () => {
@@ -47,6 +47,6 @@ describe('Trends view for patient', () => {
 
     expect(await screen.findByTestId('patient-nav-bar', {}, { timeout: 3000 })).toBeVisible()
     checkPatientNavBarAsPatient()
-    await checkPatientLayout(`${unmonitoredPatientAsTeamMember.profile.firstName} ${unmonitoredPatientAsTeamMember.profile.lastName}`)
+    await checkPatientLayout(`${patient2AsTeamMember.profile.firstName} ${patient2AsTeamMember.profile.lastName}`)
   })
 })